库存信息 add

+++ osbulkparts_system/src/main/java/cn/springboot/osbulkparts/entity/TOrderDetailInfoEntity.java
@@ -119,7 +119,7 @@ public class TOrderDetailInfoEntity implements Serializable {
     /**
      * 最小包装总量
      */
-    private Double materialMinpackageTotalamt;
+    private BigDecimal materialMinpackageTotalamt;
 
     /**
      * 未税单价
